Achievements (#5)
achievements!
this took me a couple of days, but it's finally done!
oh, and btw, there new multirow_insert extension for sqlpp, which might be useful in future
also there was a improvement in mods system, now it's support binary AND
next big feature might be a akatsuki pp system, but who knows :p

+++ src/handlers/multiplayer/room/room_change_mods_handler.cc
@@ -31,14 +31,14 @@ void shiro::handler::multiplayer::room::change_mods::handle(shiro::io::osu_packe
 
         if (match.multi_special_modes == 1) { // Free Mod
             if (match.host_id == user->user_id) {
-                if (mods & static_cast<uint32_t>(utils::mods::double_time)) {
+                if (mods & utils::mods::double_time) {
                     match.active_mods = static_cast<uint32_t>(utils::mods::double_time);
 
-                    if (mods & static_cast<uint32_t>(utils::mods::nightcore)) {
+                    if (mods & utils::mods::nightcore) {
                         match.active_mods |= static_cast<uint32_t>(utils::mods::nightcore);
                     }
                 }
-                else if (mods & static_cast<uint32_t>(utils::mods::half_time)) {
+                else if (mods & utils::mods::half_time) {
                     match.active_mods = static_cast<uint32_t>(utils::mods::half_time);
                 }
                 else {
